Downloads/x-force-bot-main/x-force-bot-main/requirements.txt
flask
langchain
langchain-google-genai
pinecone
pymongo
requests
python-dotenv
langgraph
